`ct/l1`: pull out `keyed_priority_queue` from `leveling_queue.h`
It will be used in a future commit in a new `compaction_queue.h` header.
